Change expander_dll file format. Use a null int instead of byte at end of internal files.
git-svn-id: http://crab/source/c#@49 4a3fcead-100c-0410-8434-c0d6b4fee585

+++ exe_shrink/trunk/shrinker/build.cs
@@ -61,7 +61,8 @@ class Build
 					fsIn.Close();
 				}
 			}
-			bwOut.Write((byte) 0);
+			//end of internal files null int
+			bwOut.Write(0);
 			//Use checksums?
 			if (checkSums == true)
 				bwOut.Write((byte) 1);
